third_party/WebKit/LayoutTests/platform/mac/accessibility/inline-text-bidi-bounds-for-range-expected.txt

one two אחתשתיים three four שלושהארבעה

one two אחתשתיים three four שלושהארבעה

Tests that we can compute the bounds of a range of text from the accessibility tree in bidirectional text.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Testing bounds of word one in horizontalParagraph
PASS axBounds.x is within 2 of 8
PASS axBounds.y is within 2 of 11
PASS axBounds.width is within 2 of 23
PASS axBounds.height is within 2 of 16

Testing bounds of word two in horizontalParagraph
PASS axBounds.x is within 2 of 35
PASS axBounds.y is within 2 of 11
PASS axBounds.width is within 2 of 24
PASS axBounds.height is within 2 of 16

Testing bounds of word three in horizontalParagraph
PASS axBounds.x is within 2 of 144
PASS axBounds.y is within 2 of 11
PASS axBounds.width is within 2 of 32
PASS axBounds.height is within 2 of 16

Testing bounds of word four in horizontalParagraph
PASS axBounds.x is within 2 of 180
PASS axBounds.y is within 2 of 11
PASS axBounds.width is within 2 of 26
PASS axBounds.height is within 2 of 16

Testing bounds of word אחתשתיים in horizontalParagraph
PASS axBounds.x is within 2 of 63.1015625
PASS axBounds.y is within 2 of 11
PASS axBounds.width is within 2 of 76.7421875
PASS axBounds.height is within 2 of 16

Testing bounds of word שלושהארבעה in horizontalParagraph
PASS axBounds.x is within 2 of 210.4765625
PASS axBounds.y is within 2 of 11
PASS axBounds.width is within 2 of 96.4453125
PASS axBounds.height is within 2 of 16

Testing bounds of word one in verticalParagraph
PASS axBounds.x is within 2 of 27
PASS axBounds.y is within 2 of 43
PASS axBounds.width is within 2 of 16
PASS axBounds.height is within 2 of 23

Testing bounds of word two in verticalParagraph
PASS axBounds.x is within 2 of 27
PASS axBounds.y is within 2 of 70
PASS axBounds.width is within 2 of 16
PASS axBounds.height is within 2 of 24

Testing bounds of word three in verticalParagraph
PASS axBounds.x is within 2 of 27
PASS axBounds.y is within 2 of 179
PASS axBounds.width is within 2 of 16
PASS axBounds.height is within 2 of 32

Testing bounds of word four in verticalParagraph
PASS axBounds.x is within 2 of 27
PASS axBounds.y is within 2 of 215
PASS axBounds.width is within 2 of 16
PASS axBounds.height is within 2 of 26

Testing bounds of word אחתשתיים in verticalParagraph
PASS axBounds.x is within 2 of 27
PASS axBounds.y is within 2 of 98.1015625
PASS axBounds.width is within 2 of 16
PASS axBounds.height is within 2 of 76.7421875

Testing bounds of word שלושהארבעה in verticalParagraph
PASS axBounds.x is within 2 of 27
PASS axBounds.y is within 2 of 245.4765625
PASS axBounds.width is within 2 of 16
PASS axBounds.height is within 2 of 96.4453125
PASS successfullyParsed is true

TEST COMPLETE
